c_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
absl-py=0.11.0=py36h5fab9bb_0
aiohttp=3.7.3=py36h8f6f2f9_1
argon2-cffi=20.1.0=py36h7b6447c_1
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
async_generator=1.10=py36h28b3542_0
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=py_0
bleach=3.2.1=py_0
blinker=1.4=py_1
brotlipy=0.7.0=py36h8f6f2f9_1001
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cachetools=4.2.1=pyhd8ed1ab_0
certifi=2020.12.5=py36h5fab9bb_1
cffi=1.14.4=py36hc120d54_1
chardet=3.0.4=py36h9880bd3_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.3.1=py36h0a59100_1
cycler=0.10.0=py_2
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
expat=2.2.10=he6710b0_2
fastdtw=0.2.0=py_1
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h0708190_1
gast=0.3.3=py_0
glib=2.56.2=hd408876_0
google-auth=1.24.0=pyhd3deb0d_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.35.0=py36h8e87921_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=nompi_py36ha233d17_105
hdf5=1.10.6=nompi_h6a2412b_1114
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
idna_ssl=1.1.0=py36h9f0ad1d_1001
importlib-metadata=3.4.0=py36h5fab9bb_0
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_1
jedi=0.17.2=py36_0
jinja2=2.11.2=py_0
joblib=1.0.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonschema=3.0.2=py36_0
jupyter=1.0.0=py36_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py36_0
jupyterlab_pygments=0.1.2=py_0
keras=2.4.3=py_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py36h605e78d_1
krb5=1.17.2=h926e7f8_0
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=9.3.0=hff62375_18
libgfortran5=9.3.0=hff62375_18
libgomp=9.3.0=h2828fa1_18
liblapack=3.9.0=8_openblas
libllvm10=10.0.1=he513fc3_3
libnghttp2=1.43.0=h812cca2_0
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.14.0=h780b84a_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.2.0=hdc55705_0
libuuid=1.0.3=h1bed415_2
libwebp-base=1.2.0=h7f98852_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.35.0=py36h05121d2_1
lz4-c=1.9.3=h9c3ff4c_0
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib-base=3.3.4=py36hd391965_0
mistune=0.8.4=py36h7b6447c_0
multidict=5.1.0=py36h8f6f2f9_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py36_0
nbformat=5.0.8=py_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.4.1=py_0
notebook=6.1.4=py36_0
numba=0.52.0=py36h284efc9_0
numpy=1.19.5=py36h2aa4a07_1
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1i=h7f98852_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=py_0
pandas=1.1.5=py36h284efc9_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py36_1
parso=0.7.0=py_0
patsy=0.5.1=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=8.1.0=py36h4f9996e_1
pip=21.0.1=pyhd8ed1ab_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.14.0=py36hc4f0c31_1
ptyprocess=0.6.0=py36_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyjwt=2.0.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py36h22d08a2_1
pyrsistent=0.17.3=py36h7b6447c_0
pysocks=1.7.1=py36h5fab9bb_3
python=3.6.12=hffdb5ce_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py36h8f6f2f9_0
pyzmq=19.0.2=py36he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=he28a2e2_2
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7=pyhd3deb0d_0
scikit-learn=0.23.2=py36h0573a6f_0
scipy=1.4.1=pypi_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
send2trash=1.5.0=py36_0
setuptools=49.6.0=py36h5fab9bb_3
sip=4.19.24=py36he6710b0_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.34.0=h74cdb3f_0
statsmodels=0.12.2=py36h92226af_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow=2.2.0=mkl_py36h5a57954_0
tensorflow-base=2.2.0=mkl_py36hd506778_0
tensorflow-estimator=2.2.0=pypi_0
termcolor=1.1.0=py_2
terminado=0.9.1=py36_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
tornado=6.1=py36h8f6f2f9_1
traitlets=4.3.3=py36_0
tslearn=0.5.0.5=py36h92226af_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.3=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=py36_0
wrapt=1.12.1=py36h8f6f2f9_3
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.6.3=py36h8f6f2f9_1
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.8=ha95c52a_1
